Add landing page and move editor
